Critical bug fix for SurfaceAboveGeoid.
The swapped indices in various data set not handled in this function in
two ways. Was slightly broken with default (Rampdem) dataset. Was very
broken for non-default data sets, which were probably never used by
anyone otherwise this would have been discovered way sooner.

This bug was introduced by me when switching to non-Rampdem data
in the RampdemReader data sets.