Code refactor in preparation for libAntarctica
This is a squash of many commits as I figured out how to do this
properly. I have done a minor refactor of several of these classes to
remove dependencies on EventReaderRoot in the classes that are about
to be moved.

Data modelling Antarctica has been moved to libAntarcticaRoot.

Of particular note is the GeoidModel which contains a vector which
inherits from TVector. The ellipsoid model things are encoded
there.